requirements.txt
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
pytz==2019.1
requests==2.21.0
